iconic_portal_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.0
slowapi==0.1.9
python-magic==0.4.27
fastapi-csrf-protect==0.3.4
requests==2.31.0